Enables plugin to analyze kotlin projects
