+++ plugins/de.cognicrypt.staticanalyzer/src/de/cognicrypt/staticanalyzer/results/ResultsCCUIListener.java
@@ -16,17 +16,16 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Set;
+
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IResource;
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.jdt.core.ICompilationUnit;
-import org.eclipse.jdt.core.IPackageDeclaration;
-import org.eclipse.jdt.core.JavaCore;
-import org.eclipse.jdt.core.JavaModelException;
 import org.eclipse.jface.preference.IPreferenceStore;
 import org.w3c.dom.Node;
+
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Table;
+
 import boomerang.BackwardQuery;
 import boomerang.Query;
 import boomerang.jimple.Statement;
@@ -52,11 +51,11 @@ import crypto.rules.CrySLPredicate;
 import de.cognicrypt.core.Constants;
 import de.cognicrypt.core.Constants.Severities;
 import de.cognicrypt.staticanalyzer.Activator;
+import de.cognicrypt.staticanalyzer.utilities.Utils;
 import de.cognicrypt.staticanalyzer.view.AnalysisData;
 import de.cognicrypt.staticanalyzer.view.ResultsUnit;
 import de.cognicrypt.staticanalyzer.view.StatisticsView;
 import de.cognicrypt.staticanalyzer.view.Stats;
-import de.cognicrypt.utils.Utils;
 import de.cognicrypt.utils.XMLParser;
 import soot.SootClass;
 import soot.Value;
@@ -132,15 +131,7 @@ public class ResultsCCUIListener extends CrySLAnalysisListener {
 		final int stmtId = error.hashCode();
 		HashMap<String, String> errorInfoMap = new HashMap<>();
 
-		ICompilationUnit javaFile = (ICompilationUnit) JavaCore.create(sourceFile);
-		String className = "";
-		try {
-			for (IPackageDeclaration decl : javaFile.getPackageDeclarations()) {
-				className += decl.getElementName() + ".";
-			}
-		} catch (JavaModelException e1) {
-		}
-		className += javaFile.getElementName().substring(0, javaFile.getElementName().lastIndexOf("."));
+		String className = sourceFile.getName();
 		
 		if (stat.getClassesAnalysed().containsKey(className)) {
 			AnalysisData data = stat.getClassesAnalysed().get(className);
